@@ -5,17 +5,16 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": ">=7.2",
+        "php": "^8.0",
+        "laravel/lumen-framework": "^8.0",
+        "league/flysystem": "^1.0",
         "intervention/image": "^2.5",
         "intervention/imagecache": "^2.4",
-        "laravel/lumen-framework": "^7.0",
-        "league/flysystem": "^1.0",
         "gabrielelana/byte-units": "^0.5.0"
     },
     "require-dev": {
-        "fzaninotto/faker": "^1.4",
-        "phpunit/phpunit": "^7.0",
-        "mockery/mockery": "^1.0"
+        "phpunit/phpunit": "^9.3",
+        "mockery/mockery": "^1.3.1"
     },
     "autoload": {
         "psr-4": {
