@@ -2,6 +2,8 @@
 
 namespace App\Console;
 
+use App\Console\Commands\ClientsAdd;
+use App\Console\Commands\ClientsShow;
 use Laravel\Lumen\Console\Kernel as ConsoleKernel;
 
 class Kernel extends ConsoleKernel
@@ -12,7 +14,7 @@ class Kernel extends ConsoleKernel
      * @var array
      */
     protected $commands = [
-        Commands\ClientsAdd::class,
-        Commands\ClientsShow::class,
+        ClientsAdd::class,
+        ClientsShow::class,
     ];
 }
