@@ -3,25 +3,24 @@
 namespace App\Http\Controllers;
 
 use App\File;
+use App\Http\Resources\FileResource;
 use Illuminate\Http\JsonResponse;
 use Laravel\Lumen\Routing\Controller as BaseController;
 
-class DeleteController extends BaseController
+class InfoController extends BaseController
 {
     /**
-     * Files remove.
+     * View info about file.
      *
      * @param string $client
      * @param string $fileName
      *
      * @return JsonResponse
      */
-    public function destroy(string $client, string $fileName): JsonResponse
+    public function info(string $client, string $fileName): JsonResponse
     {
         $file = File::findOrFail($client, $fileName);
 
-        $file->delete();
-
-        return response()->json(null, 204);
+        return FileResource::make($file)->response();
     }
 }
