@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\File;
 use App\Http\Resources\FileResource;
 use Exception;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Laravel\Lumen\Routing\Controller as BaseController;
@@ -14,11 +15,13 @@ class UploadController extends BaseController
     /**
      * Files upload.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param Request $request
      *
-     * @return Response
+     * @throws Exception
+     *
+     * @return JsonResponse
      */
-    public function upload(Request $request)
+    public function store(Request $request): JsonResponse
     {
         $files = [];
 
@@ -27,7 +30,7 @@ class UploadController extends BaseController
                 $file = File::save(
                     $uploaded,
                     $request->client->name,
-                    (isset($request->private) && $request->private !== 'false'),
+                    ($request->has('private') && $request->private !== 'false'),
                 );
 
                 if (!Storage::exists($file->path())) {
@@ -38,6 +41,6 @@ class UploadController extends BaseController
             });
         }
 
-        return FileResource::collection($files);
+        return FileResource::collection($files)->response();
     }
 }
