@@ -4,24 +4,21 @@ namespace App\Http\Middleware;
 
 use App\Client;
 use Closure;
+use Illuminate\Http\Request;
 
 class Authenticate
 {
     /**
      * Handle an incoming request.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \Closure  $next
+     * @param Request $request
+     * @param Closure $next
      *
      * @return mixed
      */
     public function handle($request, Closure $next)
     {
-        $request->client = Client::login($request->client, $request->header('Authorization', ''));
-
-        if (!$request->client) {
-            return abort(401);
-        }
+        $request->client = Client::loginOrFail($request->client, $request->header('x-api-key'));
 
         return $next($request);
     }
